opt-out the registry options and refactor the related code
