+++ cmd/ci-operator-checkconfig/main.go
@@ -116,7 +116,7 @@ func (o *options) loadResolver(path string) error {
 	if path == "" {
 		return nil
 	}
-	refs, chains, workflows, _, _, observers, err := load.Registry(path, load.RegistryFlag(0))
+	refs, chains, workflows, _, _, observers, err := registry.Load(path, registry.RegistryFlag(0))
 	if err != nil {
 		return err
 	}
