+++ pkg/rehearse/rehearse.go
@@ -36,7 +36,6 @@ import (
 	testimagestreamtagimportv1 "github.com/openshift/ci-tools/pkg/api/testimagestreamtagimport/v1"
 	"github.com/openshift/ci-tools/pkg/config"
 	"github.com/openshift/ci-tools/pkg/diffs"
-	"github.com/openshift/ci-tools/pkg/load"
 	"github.com/openshift/ci-tools/pkg/registry"
 )
 
@@ -253,7 +252,7 @@ func (r RehearsalConfig) createResolver(candidatePath string) (registry.Resolver
 	var observers registry.ObserverByName
 	if !r.NoRegistry {
 		var err error
-		registryRefs, chains, workflows, _, _, observers, err = load.Registry(filepath.Join(candidatePath, config.RegistryPath), load.RegistryFlag(0))
+		registryRefs, chains, workflows, _, _, observers, err = registry.Load(filepath.Join(candidatePath, config.RegistryPath), registry.RegistryFlag(0))
 		if err != nil {
 			return nil, fmt.Errorf("could not load step registry: %w", err)
 		}
@@ -396,7 +395,7 @@ func determineChangedTemplates(candidate, baseSHA, headSHA string, prNumber int,
 
 func determineChangedRegistrySteps(candidate, baseSHA string, logger *logrus.Entry) ([]registry.Node, error) {
 	var changedRegistrySteps []registry.Node
-	refs, chains, workflows, _, _, observers, err := load.Registry(filepath.Join(candidate, config.RegistryPath), load.RegistryFlag(0))
+	refs, chains, workflows, _, _, observers, err := registry.Load(filepath.Join(candidate, config.RegistryPath), registry.RegistryFlag(0))
 	if err != nil {
 		return nil, fmt.Errorf("could not load step registry: %w", err)
 	}
